requirements.txt
numpy
scikit-learn
pydflow
tensorflow==2.6.2
google
mdtraj
dpdata
